pyproject.toml
[tool.poetry]
name = "dekpy"
version = "0.1.0"
description = ""
authors = ["nuxion <nuxion@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.14.1"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
